Dedupe WrapperOpaque classes
